move Rsyslogd into helpers module
