+++ test-framework/sudo-compliance-tests/src/syslog.rs
@@ -1,56 +1,6 @@
-use sudo_test::{Child, Command, Env};
+use sudo_test::{Command, Env};
 
-use crate::{Result, SUDOERS_ALL_ALL_NOPASSWD, SUDOERS_USER_ALL_ALL, USERNAME};
-
-struct Rsyslogd<'a> {
-    _child: Child,
-    env: &'a Env,
-}
-
-impl<'a> Rsyslogd<'a> {
-    fn start(env: &'a Env) -> Result<Self> {
-        let child = Command::new("rsyslogd").arg("-n").spawn(env)?;
-        Ok(Self { _child: child, env })
-    }
-
-    /// returns the contents of `/var/auth.log`
-    fn auth_log(&self) -> Result<String> {
-        let path = "/var/log/auth.log";
-        Command::new("sh")
-            .arg("-c")
-            .arg(format!("[ ! -f {path} ] || cat {path}"))
-            .output(self.env)?
-            .stdout()
-    }
-}
-
-impl Drop for Rsyslogd<'_> {
-    fn drop(&mut self) {
-        // need to kill the daemon or `Env::drop` won't properly `stop` the docker container
-        let _ = Command::new("sh")
-            .args(["-c", "kill -9 $(pidof rsyslogd)"])
-            .output(self.env);
-    }
-}
-
-#[test]
-fn rsyslogd_works() -> Result<()> {
-    let env = Env("").build()?;
-    let rsyslog = Rsyslogd::start(&env)?;
-
-    let auth_log = rsyslog.auth_log()?;
-    assert_eq!("", auth_log);
-
-    Command::new("useradd")
-        .arg("ferris")
-        .output(&env)?
-        .assert_success()?;
-
-    let auth_log = rsyslog.auth_log()?;
-    assert_contains!(auth_log, "useradd");
-
-    Ok(())
-}
+use crate::{helpers::Rsyslogd, Result, SUDOERS_ALL_ALL_NOPASSWD, SUDOERS_USER_ALL_ALL, USERNAME};
 
 #[test]
 #[ignore = "gh421"]
